improvement(processor): improve detection if package is in a module
Improve code layout for reading and add stricter detection if a package is in a module by checking each path part.

Issue: GH-30
