+++ .github/workflows/go.yml
@@ -26,11 +26,11 @@ jobs:
     name: Build
     runs-on: ubuntu-latest
     steps:
+    - uses: actions/checkout@v3
+
     - uses: actions/setup-go@v3
       with:
-        go-version: 1.17
-
-    - uses: actions/checkout@v3
+        go-version: stable
 
     - name: Build
       run: make build
