+++ .github/workflows/golangci-lint.yml
@@ -28,11 +28,11 @@ jobs:
     name: lint
     runs-on: ubuntu-latest
     steps:
+    - uses: actions/checkout@v3
+
     - uses: actions/setup-go@v3
       with:
-        go-version: 1.17
-
-    - uses: actions/checkout@v3
+        go-version: stable
 
     - name: golangci-lint
       uses: golangci/golangci-lint-action@v3
