+++ .goreleaser.yml
@@ -3,12 +3,12 @@ builds:
   env:
   - CGO_ENABLED=0
 archives:
-- replacements:
-    darwin: Darwin
-    linux: Linux
-    windows: Windows
-    386: i386
-    amd64: x86_64
+- name_template: >-
+    {{ .ProjectName }}_
+    {{- title .Os }}_
+    {{- if eq .Arch "amd64" }}x86_64
+    {{- else if eq .Arch "386" }}i386
+    {{- else }}{{ .Arch }}{{ end }}
 checksum:
   name_template: 'checksums.txt'
 dockers:
@@ -21,7 +21,6 @@ dockers:
   dockerfile: Dockerfile.goreleaser
   build_flag_templates:
   - "--pull"
-  - "--build-arg=gomodguard_VERSION={{.Version}}"
   - "--label=org.opencontainers.image.created={{.Date}}"
   - "--label=org.opencontainers.image.name={{.ProjectName}}"
   - "--label=org.opencontainers.image.revision={{.FullCommit}}"
