+++ Dockerfile.goreleaser
@@ -1,9 +1,5 @@
-ARG GO_VERSION=1.14.2
-ARG ALPINE_VERSION=3.11
-ARG gomodguard_VERSION=
-
 # ---- App container
-FROM golang:${GO_VERSION}-alpine${ALPINE_VERSION}
+FROM golang:alpine
 WORKDIR /
 RUN apk --no-cache add ca-certificates
 COPY gomodguard /gomodguard
