+++ _example/.gomodguard.yaml
@@ -11,7 +11,7 @@ blocked:
   modules:                                                      # List of blocked modules
     - github.com/uudashr/go-module:                             # Blocked module
         recommendations:                                        # Recommended modules that should be used instead (Optional)
-          - golang.org/x/mod                           
+          - golang.org/x/mod
         reason: "`mod` is the official go.mod parser library."  # Reason why the recommended module should be used (Optional)
     - github.com/gofrs/uuid:
         recommendations:
