+++ blocked_test.go
@@ -1,42 +1,12 @@
-// nolint:scopelint
 package gomodguard_test
 
 import (
-	"fmt"
-	"os"
 	"reflect"
 	"strings"
 	"testing"
 
 	"github.com/ryancurrah/gomodguard"
 )
-
-var (
-	config *gomodguard.Configuration
-	cwd    string
-)
-
-func TestMain(m *testing.M) {
-	err := os.Chdir("_example")
-	if err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
-
-	cwd, err = os.Getwd()
-	if err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
-
-	config, err = gomodguard.GetConfig(".gomodguard.yaml")
-	if err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
-
-	os.Exit(m.Run())
-}
 
 func TestBlockedModuleIsAllowed(t *testing.T) {
 	var tests = []struct {
@@ -282,158 +252,6 @@ func TestBlockedModulesGetBlockedModule(t *testing.T) {
 			if blockedModule.IsCurrentModuleARecommendation(tt.currentModuleName) != tt.wantIsAllowed {
 				t.Errorf("got '%+v' want '%+v'", blockedModule.IsCurrentModuleARecommendation(tt.currentModuleName),
 					tt.wantIsAllowed)
-			}
-		})
-	}
-}
-
-func TestAllowedIsAllowedModule(t *testing.T) {
-	var tests = []struct {
-		testName            string
-		allowedModules      gomodguard.Allowed
-		lintedModuleName    string
-		wantIsAllowedModule bool
-	}{
-		{
-			"module is allowed",
-			gomodguard.Allowed{Modules: []string{"github.com/someallowed/module"}},
-			"github.com/someallowed/module",
-			true,
-		},
-		{
-			"module not allowed",
-			gomodguard.Allowed{},
-			"github.com/someblocked/module",
-			false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.testName, func(t *testing.T) {
-			isAllowedModule := tt.allowedModules.IsAllowedModule(tt.lintedModuleName)
-			if !reflect.DeepEqual(isAllowedModule, tt.wantIsAllowedModule) {
-				t.Errorf("got '%v' want '%v'", isAllowedModule, tt.wantIsAllowedModule)
-			}
-		})
-	}
-}
-
-func TestAllowedIsAllowedModuleDomain(t *testing.T) {
-	var tests = []struct {
-		testName                  string
-		allowedModules            gomodguard.Allowed
-		lintedModuleName          string
-		wantIsAllowedModuleDomain bool
-	}{
-		{
-			"module is allowed",
-			gomodguard.Allowed{Domains: []string{"github.com"}},
-			"github.com/someallowed/module",
-			true,
-		},
-		{
-			"module not allowed",
-			gomodguard.Allowed{},
-			"github.com/someblocked/module",
-			false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.testName, func(t *testing.T) {
-			isAllowedModuleDomain := tt.allowedModules.IsAllowedModuleDomain(tt.lintedModuleName)
-			if !reflect.DeepEqual(isAllowedModuleDomain, tt.wantIsAllowedModuleDomain) {
-				t.Errorf("got '%v' want '%v'", isAllowedModuleDomain, tt.wantIsAllowedModuleDomain)
-			}
-		})
-	}
-}
-
-func TestResultString(t *testing.T) {
-	var tests = []struct {
-		testName   string
-		result     gomodguard.Issue
-		wantString string
-	}{
-		{
-			"reason lint failed",
-			gomodguard.Issue{FileName: "test.go", LineNumber: 1, Reason: "Some reason."},
-			"test.go:1:1 Some reason.",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.testName, func(t *testing.T) {
-			result := tt.result.String()
-			if !strings.EqualFold(result, tt.wantString) {
-				t.Errorf("got '%s' want '%s'", result, tt.wantString)
-			}
-		})
-	}
-}
-
-func TestProcessorNewProcessor(t *testing.T) {
-	_, err := gomodguard.NewProcessor(config)
-	if err != nil {
-		t.Error(err)
-	}
-}
-
-func TestProcessorProcessFiles(t *testing.T) {
-	processor, err := gomodguard.NewProcessor(config)
-	if err != nil {
-		t.Error(err)
-	}
-
-	filteredFiles := gomodguard.GetFilteredFiles(cwd, false, []string{"./..."})
-
-	var tests = []struct {
-		testName   string
-		processor  gomodguard.Processor
-		wantReason string
-	}{
-		{
-			"module blocked because of recommendation",
-			gomodguard.Processor{Config: config, Modfile: processor.Modfile},
-			"blocked_example.go:9:1 import of package `github.com/uudashr/go-module` is blocked because the " +
-				"module is in the blocked modules list. `golang.org/x/mod` is a recommended module. `mod` " +
-				"is the official go.mod parser library.",
-		},
-		{
-			"module blocked because of version constraint",
-			gomodguard.Processor{Config: config, Modfile: processor.Modfile},
-			"blocked_example.go:7:1 import of package `github.com/mitchellh/go-homedir` is blocked because " +
-				"the module is in the blocked modules list. version `v1.1.0` is blocked because it does not " +
-				"meet the version constraint `<= 1.1.0`. testing if blocked version constraint works.",
-		},
-		{
-			"module blocked because of local replace directive",
-			gomodguard.Processor{Config: config, Modfile: processor.Modfile},
-			"blocked_example.go:8:1 import of package `github.com/ryancurrah/gomodguard` is blocked because " +
-				"the module has a local replace directive.",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.testName, func(t *testing.T) {
-			tt.processor.SetBlockedModules()
-			results := tt.processor.ProcessFiles(filteredFiles)
-			if len(results) == 0 {
-				t.Fatal("result should be greater than zero")
-			}
-
-			foundWantReason := false
-			allReasons := make([]string, 0, len(results))
-			for _, result := range results {
-				allReasons = append(allReasons, result.String())
-
-				if strings.EqualFold(result.String(), tt.wantReason) {
-					foundWantReason = true
-				}
-			}
-
-			if !foundWantReason {
-				t.Errorf("got '%+v' want '%s'", allReasons, tt.wantReason)
 			}
 		})
 	}
