+++ tools.go
@@ -1,0 +1,5 @@
+//go:build tools
+
+package gomodguard
+
+import _ "github.com/t-yuki/gocover-cobertura"
